+++ pyproject.toml
@@ -47,6 +47,7 @@ classifiers = [
 dependencies = ["databricks-sdk~=0.30",
                 "databricks-labs-lsql>=0.5,<0.11",
                 "databricks-labs-blueprint>=0.8,<0.9",
+                "databricks-labs-pytester>=0.2",
                 "PyYAML>=6.0.0,<7.0.0",
                 "sqlglot>=25.5.0,<25.22",
                 "astroid>=3.2.2"]
@@ -77,7 +78,7 @@ dependencies = [
     "pylint~=3.2.2",
     "pylint-pytest==2.0.0a0",
     "databricks-labs-pylint~=0.4.0",
-    "pytest~=8.1.0",
+    "pytest~=8.3.3",
     "pytest-cov~=4.1.0",
     "pytest-mock~=3.14.0",
     "pytest-timeout~=2.3.1",
